Refactoring (#145)
